+++ src/main/resources/test-files/test.svelte
@@ -1,0 +1,16 @@
+<script>
+  import { onMount } from "svelte"
+
+  export let foo
+
+  const bar = 42
+  let state = ""
+
+  onMount(() => {
+    state = "mounted"
+  })
+
+  const handleClick = () => {
+    state = "clicked"
+  }
+</script>