+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ## [Unreleased]
 
+## [1.0.83]
+- Strings' content background color in BG heme stays the same when the line is highlighted by IDE, previously background changed to line highlight color.
+- String's quotation marks' background color in dynamic now and changes to a line highlight color, previously it was always of a default background color.
+- String quotations marks rendering speed improvement. 
+- Better Scala support: strings, constants.
+
 ## [1.0.82]
 - Updated changelog for previous versions.